+++ docs/handbook/sources/build.sh
@@ -79,7 +79,7 @@ done
 rm -f temp0
 
 sed '1,19d' ../../../library/NOTES.md | pandoc -f gfm -t rst -o libraries/overview.rst
-sed '1,19d' ../../../library/ada_boost/NOTES.md | pandoc -f gfm -t rst -o libraries/ada_boost.rst
+sed '1,19d' ../../../library/adaptive_boosting_classifier/NOTES.md | pandoc -f gfm -t rst -o libraries/adaptive_boosting_classifier.rst
 sed '1,19d' ../../../library/agglomerative/NOTES.md | pandoc -f gfm -t rst -o libraries/agglomerative.rst
 sed '1,19d' ../../../library/amqp/NOTES.md | pandoc -f gfm -t rst -o libraries/amqp.rst
 sed '1,19d' ../../../library/anomaly_detection_protocols/NOTES.md | pandoc -f gfm -t rst -o libraries/anomaly_detection_protocols.rst